Update D3 library from 3.3.13 to 3.4.13.
Version upgrade of the D3 library to version 3.4, specifically
version 3.4.13, to grab minor fixes. 3.3.13 was the last version
before 3.4, and no major breaking changes seemed to occur.

The need here is support for localization of dates, as 3.3.13
hard-codes English version of dates and times in the javascript,
where this was removed as of 3.4.0.