Add field heuristics
This attempts to make a best guess at which fields to use for:
- Initial filter fields
- Title and subtitle on popup/modal
- Image / video
- Which field to color
If any of these values are set on `state` then that value is preferred. This should render something 'good enough' to start off, so the user does not need to configure any fields to start exploring data.