Handle features without unique ids
Previously the code depended on each geoJSON feature having a unique feature.id field. This change checks first whether those ids exist and they are actually unique, and if not looks for any unique features.properties and if not then generates a unique id. The id is needed for linking to a feature through the UI interaction moving through the app, in particular opening a feature detail view.