Fix browserify transform for required module
Browserify respects the `browserify: transform` field of `package.json` for required modules. When babelify was listed here it tried to babelify es5 code when this module is required from npm, which caused an error. Also having transforms in two places (as command line options and as package.json options) was confusing, this moves them all to cli options.